+++ plugins/modules/oci_analytics_instance_facts.py
@@ -186,6 +186,36 @@ analytics_instances:
                             returned: on success
                             type: str
                             sample: description_example
+                private_source_scan_hosts:
+                    description:
+                        - List of Private Source DB SCAN hosts registered with Private Access Channel for access from Analytics Instance.
+                    returned: on success
+                    type: complex
+                    contains:
+                        scan_hostname:
+                            description:
+                                - "Private Source Scan hostname. Ex: db01-scan.corp.example.com, prd-db01-scan.mycompany.com."
+                            returned: on success
+                            type: str
+                            sample: scan_hostname_example
+                        scan_port:
+                            description:
+                                - Private Source Scan host port. This is the source port where SCAN protocol will get connected (e.g. 1521).
+                            returned: on success
+                            type: int
+                            sample: 56
+                        description:
+                            description:
+                                - Description of private source scan host zone.
+                            returned: on success
+                            type: str
+                            sample: description_example
+                network_security_group_ids:
+                    description:
+                        - Network Security Group OCIDs for an Analytics instance.
+                    returned: on success
+                    type: list
+                    sample: []
         vanity_url_details:
             description:
                 - Map of VanityUrl unique identifier key as KEY and VanityUrl Object as VALUE.
@@ -335,6 +365,12 @@ analytics_instances:
                     returned: on success
                     type: str
                     sample: "ocid1.subnet.oc1..xxxxxxEXAMPLExxxxxx"
+                network_security_group_ids:
+                    description:
+                        - Network Security Group OCIDs for an Analytics instance.
+                    returned: on success
+                    type: list
+                    sample: []
                 network_endpoint_type:
                     description:
                         - The type of network endpoint.
@@ -343,7 +379,7 @@ analytics_instances:
                     sample: PUBLIC
                 whitelisted_ips:
                     description:
-                        - Source IP addresses or IP address ranges igress rules.
+                        - Source IP addresses or IP address ranges in ingress rules.
                     returned: on success
                     type: list
                     sample: []
@@ -361,10 +397,16 @@ analytics_instances:
                             sample: "ocid1.resource.oc1..xxxxxxEXAMPLExxxxxx"
                         whitelisted_ips:
                             description:
-                                - Source IP addresses or IP address ranges igress rules.
+                                - Source IP addresses or IP address ranges in ingress rules.
                             returned: on success
                             type: list
                             sample: []
+                whitelisted_services:
+                    description:
+                        - Oracle Cloud Services that are allowed to access this Analytics instance.
+                    returned: on success
+                    type: list
+                    sample: []
         service_url:
             description:
                 - URL of the Analytics service.
@@ -397,7 +439,13 @@ analytics_instances:
             "private_source_dns_zones": [{
                 "dns_zone": "dns_zone_example",
                 "description": "description_example"
-            }]
+            }],
+            "private_source_scan_hosts": [{
+                "scan_hostname": "scan_hostname_example",
+                "scan_port": 56,
+                "description": "description_example"
+            }],
+            "network_security_group_ids": []
         },
         "vanity_url_details": {
             "key": "key_example",
@@ -424,12 +472,14 @@ analytics_instances:
         "network_endpoint_details": {
             "vcn_id": "ocid1.vcn.oc1..xxxxxxEXAMPLExxxxxx",
             "subnet_id": "ocid1.subnet.oc1..xxxxxxEXAMPLExxxxxx",
+            "network_security_group_ids": [],
             "network_endpoint_type": "PUBLIC",
             "whitelisted_ips": [],
             "whitelisted_vcns": [{
                 "id": "ocid1.resource.oc1..xxxxxxEXAMPLExxxxxx",
                 "whitelisted_ips": []
-            }]
+            }],
+            "whitelisted_services": []
         },
         "service_url": "service_url_example",
         "time_created": "2013-10-20T19:20:30+01:00",